server-requirements.txt
pymongo
pyYAML
pika
flask
pyjwt
requests